requirements.txt
cryptography==41.0.3
PyJWT==2.8.0
pycryptodome==3.18.0
requests==2.31.0
rsa==4.9